requirements.txt
pandas 
numpy 
scikit-learn
matplotlib
fastapi
uvicorn
pydantic
PyYAML
python-box[all]~=7.0
PyPDF2
python-docx
nltk
tqdm
sentence-transformers
faiss-cpu
transformers










